requirements.txt
openai
python-dotenv
ipykernel
jupyter
pylint